Further work
I removed ReloadLevelObjects as it is just a less efficient version of LoadLevelObjects.

Moved the Renderer Render call to the CoreEngine Render right after the level render is called.

moved the mesh render to a new function called render in levelGraph. (this has the benifit of removing the
need to continuesly call GetInstance for the loop)

Removed superfiouly call in Eventhandler for window resizing. left it in CoreEngine.

Moved Renderer initialization to CoreEngine Init.